requirements.txt
discord
requests
pdfminer.six
pdf2image
PyNaCl
google-generativeai
pandas
openpyxl